Move Task class into Job #7
Remove Task.java file. This inner Task class can potentially be removed if we can change the data structure that holds theMachine and theTime. This could be done with a Pair<>() object potentially.
