+++ src/applications/Job.java
@@ -54,4 +54,24 @@ class Job {
         return id;
     }
 
+    // top-level nested classes
+    private class Task {
+        // data members
+        private int machine;
+        private int time;
+
+        // constructor
+        Task(int theMachine, int theTime) {
+            machine = theMachine;
+            time = theTime;
+        }
+
+        public int getMachine() {
+            return machine;
+        }
+
+        public int getTime() {
+            return time;
+        }
+    }
 }
